// framework/glsp-client/tasklist-glsp/src/framework/common/types/types.ts
export type GlspServiceInjectors = {
    BatchValidationHandler?: (services: GlspLangiumSharedServices) => BatchValidationHandler;
    DocumentAstHandler?: (services: GlspLangiumSharedServices) => DocumentAstHandler;
    ScopingInformationListener?: (services: GlspLangiumSharedServices) => ScopingInformationListener;
    BatchValidationListener?: (services: GlspLangiumSharedServices) => BatchValidationListener;
    GlspConnection?: (services: GlspLangiumSharedServices) => GlspConnection;
    ExternalNodeInformationService?: (services: GlspLangiumSharedServices) => ExternalNodeInformationService;
};

/** GLSP-Langium integration services */
export type GlspServices = {
    BatchValidationHandler: BatchValidationHandler;
    DocumentAstHandler: DocumentAstHandler;
    ScopingInformationListener: ScopingInformationListener;
    BatchValidationListener: BatchValidationListener;
    GlspConnection: GlspConnection;
    ExternalNodeInformationService: ExternalNodeInformationService;
};

/** Additional shared Langium services */
export type GlspLangiumSharedGlspServices = {
    glsp: GlspServices;
};

/** Langium shared services extended with {@link GlspServices} */
export type GlspLangiumSharedServices = LangiumSharedServices & GlspLangiumSharedGlspServices;